refactor(github): collocate service with route handler
