+++ apps/github/app/api/users/cron/schedule-sync-jobs/route.ts
@@ -1,5 +1,5 @@
 import { NextResponse } from 'next/server';
-import { scheduleSyncJobs } from '../../../../services/scheduleSyncJobs.service';
+import { scheduleSyncJobs } from './service';
 
 export async function GET() {
   const result = await scheduleSyncJobs('users');
